wpa_supplicant needs the same bindist as openssl.
At some point between November 2018 and February 2019, wpa_supplicant
started taking the same bindist USE flag as openssl.  This is probably a
good thing, since I could get away with not forcing bindist on (since I am
in fact making binary packages), but for now I'll just add
wpa_supplicant to the list of packages that get bindist turned on as my
understanding is that that flag controls EC cryptography in openssl.